Refactoring Parsers and fixing token locations
